connected index and lists file into one
